requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
pyotp==2.9.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
